rewrite site generator
This commit introduces breaking code. It is a complete rewrite of site
generator. Generator prebuilds the site by fetching (and transforming into
html) data from the meetup server and converting markdown files to html.
Additionally, it bundles javascript that needs to be executed and
appends it to corresponding html pages.